scheduler: finishing scheduler items.
- Cleaning up old secrets after their frontend Token CRs have been
  removed.
- Cleaning up old services after their backend STS or STS pods/replica
  counts have been removed.
- Re-creating frontend Services for Stream STS & STS pods when needed.
- Scheduler is now creating frontend Services for Stream STS as part of
  the StatefulSetUpdated handler. This is where Ingress creation will go
  in the future as well. Also deleting the STS if the Stream doesn't
  exist.
- Cleaning up old services when an STS is deleted.

+++ hadron-core/src/crd/stream.rs
@@ -41,7 +41,7 @@ pub struct StreamSpec {
     /// This value can be dynamically scaled up and down and directly corresponds to the number of
     /// pods in the corresponding StatefulSet. Scaling down the number of partitions
     /// will cause the data of the removed partitions to be lost. Use with care.
-    pub partitions: u8,
+    pub partitions: u32,
     /// An optional TTL in seconds specifying how long records are to be kept on the stream.
     ///
     /// If `0`, then records will stay on the stream forever. Modifying this value will require a
